It is now possible to have list variables and to expand them in list context or string context.  List context list expansions use the same @ form introduced last week.  String context list expansions cause the list to be joined in a shell-compatible way by subprocess.list2cmdline.
The sample v8.gyp is updated to use list variable expansions in both list and
string context, exercising the above change and simplifying the "actions"
sections substantially.

Remove the need for separate *_SCRIPT variables.  This was only needed because
Xcode variables are represented as $(...) but in scripts, shell environment
variable representation must be used, such as ${...}.  The Xcode generator now
looks for Xcode-style variable references and replaces them with shell-style
ones.  This makes it cumbersome to use $() in actions, but that's not really
portable and it's probably OK to advise people not to do too much on command
lines.  If something more involved is needed, put it in a script that's called
by the action as its command.

The Xcode generator now also runs scripts via "exec".