mac: Tests for LD_DYLIB_INSTALL_NAME and DYLIB_INSTALL_NAME_BASE.
Also fix a bug uncovered by this: Xcode provides a
default -install_name. (It looks like the Content Shell postbuilds
depend on this default, but I discovered that only after writing
this CL.)

Fix another bug: ld errors out if -install_name is passed to a
non-shared library, and Xcode doesn't pass it on in that case.
Mimic that behavior.

BUG=none
TEST=unit tests. `ninja content_shell` builds functional content shell
on mac.
Review URL: https://chromiumcodereview.appspot.com/9956006