ninja windows: workaround for too long names
Hash rule_name/unique_name for now to make the path shorter and get builds working again. We should come up with a better way of doing this in general.

(Jay was running into this problem, and found this to be a reasonable/working workaround)

BUG=276
TEST=

Review URL: https://chromiumcodereview.appspot.com/10665008