Fix how quotes are rendered in ninja and msvs under cygwin.
Specifically, action strings would render like:
   'action': ['python', '-c', 'print \'Hello world!\'']  ->
     bash -c "'python' '-c' 'print \'Hello world!\''" ->
       python -c 'print 'Hello world!'' ->
         print Hello world!

   'action': ['python', '-c', 'print "Hello world"] ->
     bash -c "'python' '-c' 'print "Hello world!"'" ->
       python -c 'print Hello world!' ->
         print Hello world!

Now they'll render as:
   'action': ['python', '-c', 'print \'Hello world!\'']  ->
     bash -c "'python' '-c' 'print '\''Hello world!'\'''" ->
       python -c 'print '\''Hello world!'\''' ->
         print 'Hello world'!

   'action': ['python', '-c', 'print "Hello world"] ->
     bash -c "'python' '-c' 'print \"Hello world!\"'" ->
       python -c 'print \"Hello world!\"' ->
         print "Hello world!"

This will preserve the intent of the original command.

Msvs was failing on the double-quotes test case because contained
quotes were not properly escaped when constructing the
cmd-runs-bash-runs-command line.

There's probably a higher-level bug here which is that quoting needs
to be done in a deterministic (and common) way across all the
generators.

BUG=132157

Review URL: https://chromiumcodereview.appspot.com/10782040