ninja windows: don't use /allowisolation with a .def file
Apparently (per http://build.chromium.org/f/client/gyp/) using a .def file
makes the linker think it's linking a DLL when it's actually an exe, which
means it errors out due to the just-added /ALLOWISOLATION flag for manifests.
So, if we have a def file, don't add that flag. This appears to match what the
IDE does implicitly.

TBR=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10827024