ninja/mac: Fix a regression from r1709.
r1709 set the output of a target to the stamp file if the target has no
linkable object files.  That's ok, but Target.Linkable() still considered
the target to be linkable, so the linker then tried to link in the stamp
file, which made ld complain (see e.g.
http://build.chromium.org/p/tryserver.chromium/builders/mac/builds/82504/steps/compile/logs/stdio
).  As a fix, set the target type to None if it doesn't have a binary, the
Linkable() will return False.

R=mark@chromium.org

Review URL: https://codereview.chromium.org/23877003