ninja: Stop lower-casing outputs on case-insensitive file systems.
This was added to fix a problem with file renames that rename a file only
in captialization (https://chromiumcodereview.appspot.com/10907140).  Now
that we use depsmode (which reads deps from a flat text file instead of
doing a lookup for a depsfile on the file system, and which always does
so case-sensitively), this is no longer necessary.  The test we added
when we added this hack still passes.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/23542007