Add a flag to enable the fix for r1717
It seems like Chrome (and potentially other projects using gyp) depend on the
bug that was fixed in r1717.  So for now, only enable the fix if the target has
'allow_sharedlib_linksettings_propagation' explicitly set to False (by default,
it is True).

This will allow Chrome and other projects to continue building until they are
fixed to not depend on this bug, at which point, we could remove the flag.

BUG=
R=mark@chromium.org

Review URL: https://codereview.chromium.org/23557010