win ninja: Speculative fix for pdb locking errors
There's longstanding occasional errors where the compiler complains that
the pdb is locked. This isn't really a condition that should be exposed
to the build system, and no one has been able to come up with a reliable
repro. Retrying causes the build to succeed.

In chromium, two targets that often fail this way are 'net' and
'libphonenumber_without_metadata'. These targets are distinctive in that
they build early in the build, contain at least one .c file, and use
pch for their .cc files.

The locking error seems to manifest with the compile step for the obj
associated with creating the .pdb, so the hypothesis is that the .c
compilation is colliding with the creation of the pdb that the
precompile does. This race should be handled by mspdbsrv, but for
appears to be buggy for whatever reason. (This bug has been filed on
connect.microsoft.com, but without a repro it's not very likely that
action will be taken.)

In order to avoid this, have the .c and .cc files use separate .pdb
files. This is otherwise unnecessary, but shouldn't cause any material
drawback either.

R=thakis@chromium.org
BUG=chromium:142362

Review URL: https://codereview.chromium.org/88303004