win: Add NoImportLibrary flag for ninja generator
This adds a flag that tells ninja not to expect an import library
to be generated. This is only needed very rarely in the case of
DLLs that have no exported symbols.

There is no corresponding functionality in the msvs generator, so
the test and functionality are ninja generator-only.

R=grt@chromium.org
BUG=gyp:397, chromium:451499

Review URL: https://codereview.chromium.org/890773004